+++ src/engine/galaxy/common/ai/platform/CPlatformVertical.cpp
@@ -129,9 +129,8 @@ void CPlatformVertical::draw()
     
     if( scrx < gameres.w && scry < gameres.h && exists )
     {
-        Uint16 showX = scrx+Sprite.getXOffset();
-        Uint16 showY = scry+Sprite.getYOffset();
-
+        int showX = scrx+Sprite.getXOffset();
+        int showY = scry+Sprite.getYOffset();
 
         if(drawFire && dimFire)
         {
